+++ CHANGELOG.md
@@ -7,6 +7,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: `DeskLayout`: Add layout component.
 - Feature: `DeskMenu`: Add organism component.
 
+## [6.7.0] - 2021-12-08
+
+Features:
+
+  - `RadioButton`: Add `small` to the size prop.
+  - `RadioButtonSet`: Add `small` to the size prop.
+  - `ContentCard`: Add component.
+  - `TextButton`: Add component.
+  - `AirplaneIconNext`: Add component.
+  - `BubbleIconNext`: Add component.
+
 ## [6.6.0] - 2021-12-02
 
 Features